SilhouetteMethodDemonstration
Given a data set, runs the Silhouette Method to determine the optimal number of clusters. It does so by running k-means clustering for various k values and finding the silhouette coefficient, which ranges from [-1,1]. A score close to 1 indicates that the point  is close to the center of its assigned cluster. A score of 0 indicated the point is directly in between two clusters. A score less than 0 indicates it is closer to the center of a different cluster and is therefore probably in the wrong cluster. It also tracks the amount of the points in each cluster. The k value which has the best combination of a high average overall silhouette score and evenly distributed states is said to be the optimal k value. The program then displays the two plots for each possible k value. The first plot displays the average overall silhouette score and how the each cluster's score compares to that and the amount of points in each cluster. The second plot displays the data sorted by k means with the associated k value and labels each group with its corresponding cluster number. It outputs the overall average silhouette score as well.